+++ jigoshop.php
@@ -182,6 +182,30 @@ if (is_admin()) {
 	register_activation_hook(__FILE__, 'install_jigoshop');
 }
 
+
+function jigoshop_admin_footer($text) {
+	$screen = get_current_screen();
+
+	if (strpos($screen->base, 'jigoshop') === false && strpos($screen->parent_base, 'jigoshop') === false && !in_array($screen->post_type, array('product', 'shop_order'))) {
+		return $text;
+	}
+
+	return sprintf(
+		'<a target="_blank" href="https://www.jigoshop.com/support/">%s</a> | %s',
+		__('Contact support', 'jigoshop'),
+		str_replace(
+			array('[stars]','[link]','[/link]'),
+			array(
+				'<a target="_blank" href="http://wordpress.org/support/view/plugin-reviews/jigoshop#postform" >&#9733;&#9733;&#9733;&#9733;&#9733;</a>',
+				'<a target="_blank" href="http://wordpress.org/support/view/plugin-reviews/jigoshop#postform" >',
+				'</a>'
+			),
+			__('Add your [stars] on [link]wordpress.org[/link] and keep this plugin essentially free.',WYSIJA)
+		)
+	);
+}
+add_filter('admin_footer_text', 'jigoshop_admin_footer');
+
 /**
  * Adds Jigoshop items to admin bar.
  */